Improve log

+++ In-Home Audio/Streamer.cpp
@@ -153,7 +153,7 @@ SpeakerStreamer::SpeakerStreamer() : m_NumUsed(0) {
     
 }
 void SpeakerStreamer::onGetStats(sf::Uint8 ChannelCount,sf::Uint32 SampleRate) {
-    std::cout<<"Stats "<<ChannelCount<<" "<<SampleRate<<std::endl;
+    std::cout<<"Stats "<<(int)ChannelCount<<" "<<(int)SampleRate<<std::endl;
     initialize(ChannelCount,SampleRate);
     play();
 }
